Refactoring Переместил context.py в папку schemas
