+++ bot.py
@@ -24,6 +24,7 @@ def bot():
     bot_logger = load_log_config()
     bot_logger.info('Инициализация бота')
 
+    # TODO Предусмотреть создание папки, на случай, ее нет
     persistence = PicklePersistence(filepath="db/conversationbot")
     application = (
         Application.builder()
